Renamed disc/disc.py to discord/discord.py. Directory and file modified. Removed disc endpoint from main.py, changed import.
